Testing armor and weapons for annoying accidental fights
Adds a couple stuff  that isnt going to stay in the mod as it is. Everything will most likely be reworked to be something serious in a later time period. For now, these items are here because of the amount of times ive summoned snowrium outside of the snow biome (you cant butcher it either)